README.md: Minor tweaks
I think backticks \`\` in Markdown should only be used for things
that one would type at a terminal, basically.   So `Ignition`
isn't right because one would never type it as a command or argument.
And in this case it's more useful to link to the project Github page.

Also link to fcct while we're here.